apt: new packet
this is an attempt to correct mistakes in the bionic packet

 # keep things separate

stand alone repo keep bionic and GNU separate. if you want nothing to do with bionic why force it down someone's throat. this contains only the glibc list no bionic packets. if a dependency is missing add it to the glibc list don't mix different C libs it just makes a mess.

there are probably only ten bionic packets like ca-certificates/stable  that is missing. now that I checked even that was added as ca-certificates-glibc/glibc. if there is a name conflict for those that insist on mixing both lists the suit can be appended to differentiate abc/stable for bionic and abc/glibc for glibc

the -glibc suffix which is called a prefix in the misguided build script can eventually be removed but this is beyond my education duties.

I don't fundamentally understand how the global vision failed and became so myopic. it is brilliant in the end Debian and termux packets will be byte identical . proot and glibc should have been on by default that's probably what created this mess to begin with. hardly anyone that first use termux understand for a long time that the defaults are wrong essentially . I am only beginning to realise it after wasting so much time on meaningless patches.

I am not fixing anything that works in proot glibc anymore  have already done that for years for no purpose

 # no android paths in the source code

use  settings and wrappers instead of hard paths. just makes me cringe when I see a gigantic android path in the source code . some of us always use proot an have no need for any of it. there is a draft apt.conf for non proot but I don't use it and have not completed it. I can't spend my entire life on circumventing problems that proot solve in one swat. it takes one second to start proot from .bashrc I don't know any reason to not use it

 # to-do remove root path prefix from packet archive

this remains to be solved and needs a education  effort that's beyond my
energy I am already exhausted with a managing role in United nations for
south Africa

use proot or dpkg --root=abc to install packets and not the archive structure. I use another terminal and have to proot to an insane path . unbelievable
how this was allowed to happen. this is where my files are installed because of this mistake

~~~
/data/data/open.term <- my terminal
/data/data/com.termux/files/usr/bin/abc <- path prefix in abc.deb!
/data/data/open.term/data/data/com.termux/files/usr/bin/abc <- this is insane
~~~

that prefix should not be in the archive. with proot nothing has to be
specified everything works already. if some one refuse to use proot for
inexplicable reason the correct way is dpkg --root=/data/data/abc/abc.
absolutely not in the archive destroying it for all sane purposes

~~~
/usr/bin/abc <- this is the correct path for the abc.deb archive for proot or dpkg --root=/data/data/com.termux/files/usr
~~~

as a side note I don't even use these paths in termux. when I first installed it I linked them to a shorter path

~~~
files/usr -> /data/data/com.termux/bionic
files/usr/glibc -> /data/data/com.termux/glibc
~~~

last thing I need is even longer paths

 # copy Debian packet to termux

 after endless searching I think I found the Debian build system for R. the build.sh file is called debian/control debian/rules. it's in the same repo as the app?!

~~~
https://salsa.debian.org/edd/r-base

ack "Depends" debian/control|head -1
Build-Depends: gcc (>= 4:4.9.2-2), g++ (>= 4:4.9.2-2), gfortran (>= 4:4.9.2-2), libblas-dev, liblapack-dev, tcl8.6-dev, tk8.6-dev, bison, groff-base, libncur...

head -2 debian/rules

dpkg-buildpackage -B
~~~

 so the Debian build recipes are all over the place impossible to find them in one place

this packet already works with proot but I will try to copy the rules file to build.sh as an exercise